[Refactor, Test] CommentService 내 의존성 제거 및 리팩토링 및 컨트롤러 테스트 코드 작성, 인증 관련 클래스 구조 변경 시큐리티

+++ src/main/java/com/yapp18/retrospect/service/CommentService.java
@@ -1,6 +1,7 @@
 package com.yapp18.retrospect.service;
 
 import com.yapp18.retrospect.config.ErrorInfo;
+import com.yapp18.retrospect.config.TokenErrorInfo;
 import com.yapp18.retrospect.domain.comment.Comment;
 import com.yapp18.retrospect.domain.comment.CommentRepository;
 import com.yapp18.retrospect.domain.post.Post;
@@ -13,6 +14,7 @@ import com.yapp18.retrospect.web.dto.ApiPagingResultResponse;
 import com.yapp18.retrospect.web.dto.CommentDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,81 +26,62 @@ import java.util.stream.Collectors;
 @Service
 public class CommentService {
     private final CommentRepository commentRepository;
-    private final UserRepository userRepository;
     private final PostRepository postRepository;
-    private final CommentMapper commentMapper;
 
     @Transactional
-    public Long inputComments(CommentDto.InputRequest inputRequest, Long userIdx){
-        User user = userRepository.findByUserIdx(userIdx)
-                .orElseThrow(() -> new EntityNullException(ErrorInfo.USER_NULL));
-
-        Post post = postRepository.findByPostIdx(inputRequest.getPostIdx())
-                .orElseThrow(() -> new EntityNullException(ErrorInfo.POST_NULL));
-
-        return commentRepository.save(inputRequest.toEntity(post, user)).getCommentIdx();
+    public Comment inputComments(Comment comment){
+        //타 서비스 의존성을 제거와 코드 가독성을 위해 Entity 외의 메서드 parameter는 최소화 할 것
+        return commentRepository.save(comment);
     }
 
     @Transactional(readOnly = true)
-    public ApiPagingResultResponse<CommentDto.BasicResponse> getCommmentsListByPostIdx(Long postIdx, Long cursorIdx, Long userIdx, Pageable page){
-        Post post = postRepository.findByPostIdx(postIdx)
+    public List<Comment> getCommmentsListByPostIdx(Long postIdx, Pageable page){
+        Post post = postRepository.findById(postIdx)
                 .orElseThrow(() -> new EntityNullException(ErrorInfo.POST_NULL));
 
-        List<Comment> result = cursorIdx == 0 ?
-                commentRepository.findAllByPost(post, page) :
-                commentRepository.cursorFindAllByPost(cursorIdx, postIdx, page);
-
-        Long lastIdx = result.isEmpty() ? null : result.get(result.size() - 1).getCommentIdx();
-
-        return new ApiPagingResultResponse(isNext(post, lastIdx),
-                result.stream()
-                .map(comment -> commentMapper.commentToBasicResponse(comment, isWriter(comment.getUser().getUserIdx(), userIdx)))
-                .collect(Collectors.toList()));
+        return commentRepository.findAllByPost(post, page);
     }
 
     @Transactional(readOnly = true)
     public Long getCommmentsCountByPostIdx(Long postIdx){
-        Post post = postRepository.findByPostIdx(postIdx)
-                .orElseThrow(() -> new EntityNullException(ErrorInfo.POST_NULL));
+        Post post = postRepository.findByPostIdx(postIdx);
         return commentRepository.countCommentByPost(post);
     }
 
     @Transactional(readOnly = true)
-    public CommentDto.BasicResponse getCommmentsByIdx(Long commentIdx){
+    public Comment getCommmentsByIdx(Long commentIdx){
         return commentRepository.findById(commentIdx)
-                .map(commentMapper::commentToBasicResponse)
                 .orElseThrow(() ->  new EntityNullException(ErrorInfo.COMMENT_NULL));
     }
 
 
     @Transactional
-    @PreAuthorize("#comment.user.userIdx == #userIdx")
-    public CommentDto.BasicResponse updateCommentsByIdx(CommentDto.UpdateRequest updateRequest, Comment comment, Long userIdx){
-        User user = userRepository.findByUserIdx(userIdx)
-                .orElseThrow(() -> new EntityNullException(ErrorInfo.USER_NULL));
+//    @PreAuthorize("#oldComment.user.userIdx == #user.userIdx")
+    public Comment updateComments(Comment newComment){
+        Long commentIdx = newComment.getCommentIdx();
 
-        Post post = postRepository.findByPostIdx(comment.getPost().getPostIdx())
-                .orElseThrow(() -> new EntityNullException(ErrorInfo.POST_NULL));
+        Comment oldComment = commentRepository.findById(commentIdx)
+                .orElseThrow(() -> new EntityNullException(ErrorInfo.COMMENT_NULL));
 
-        return commentMapper.commentToBasicResponse(commentRepository.save(
-                comment.update(updateRequest.getComments(), post, user)
-                )
-        );
+        if(!oldComment.isWriter(newComment.getUser())){
+            throw new AccessDeniedException(TokenErrorInfo.ACCESS_DENIED.getMessage());
+        }
+
+        oldComment.update(newComment);
+
+        return commentRepository.save(oldComment);
     }
 
     @Transactional
-    @PreAuthorize("#comment.user.userIdx == #userIdx")
-    public void deleteComments(Comment comment, Long userIdx){
-        System.out.println(comment.getUser().getUserIdx());
-        System.out.println(userIdx);
+    public void deleteComments(User user, Long commentIdx){
+        Comment comment = commentRepository.findById(commentIdx)
+                .orElseThrow(() -> new EntityNullException(ErrorInfo.COMMENT_NULL));
+
+        if(!comment.isWriter(user)){
+            throw new AccessDeniedException(TokenErrorInfo.ACCESS_DENIED.getMessage());
+        }
+
         commentRepository.delete(comment);
-    }
-
-
-    // 작성자 판별
-    private boolean isWriter(Long commentUserIdx, Long userIdx){
-        if (userIdx == 0L) return false;
-        return commentUserIdx.equals(userIdx);
     }
 
     // 다음 페이지 여부 확인
